Move constants to tests
Because they should only be used in tests, but not in production code
